requirements.txt
pandas
numpy
faker
psycopg2-binary
sqlalchemy
python-dotenv
